requirements.txt
setuptools>=65.5.1
flask==2.0.1
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
selenium==4.1.0
newspaper3k==0.2.8
beautifulsoup4==4.9.3
reportlab==4.0.4
python-docx==0.8.11
trafilatura==1.6.1
langdetect==1.0.9
APScheduler==3.10.4
webdriver-manager==4.0.0
requests==2.26.0
psutil==5.9.5
waitress==3.0.0
cloudscraper==1.2.58
undetected-chromedriver>=3.5.0
readability-lxml==0.8.1
flask_limiter==2.4.0
gunicorn==20.1.0
python-dotenv==1.0.0